requirements.txt
DMXEnttecPro
PySide6
artnet @ https://github.com/Frangitron/python-artnet @ main
dataclasses-json
dateparser
jinja2
ledboarddesktop @ git+https://github.com/Frangitron/ledboard-desktop @ main
ledboardlib @ git+https://github.com/Frangitron/ledboard-lib @main
pyside6-helpers @ git+https://github.com/MrFrangipane/pyside6-helpers @ main
python-helpers @ git+https://github.com/MrFrangipane/python-helpers @ main
python-rtmidi
